requirements.txt
requests
beautifulsoup4
pandas
pytz
python-dotenv
